feat: uint64 node IDs — shorter unique IDs (8 bytes from keccak256)
BREAKING: Node IDs changed from uint256 (78 digits) to uint64 (18 digits)
- ComputeRegistry: uint64(bytes8(keccak256(...))) instead of full uint256
- ComputeMarketplace + ComputeIndexToken: updated interface to uint64
- Redeployed all contracts (v5) to testnet
- CLI: updated ABI types + all commands work with uint64 IDs
- Rust agent: updated addresses, rebuilt successfully

New contract addresses (testnet chainId 968):
  ComputeRegistry:    0x8b68ae929A0Cbe32F6F0121881B42Ef9D9213eB5
  PriceOracle:        0x2BF8219f6b296A85904e4A486963496c3A0d1b43
  ComputeMarketplace: 0x89b6fBFB647B8a07c4d1520871440f0B01314f87
  ComputeIndexToken:  0x11D29Bf60E75f3A3Dc3b46fC7dfaafc5BdB6825E

Example: node-id:853229305740651737 (was 78 digits before)
Tested: setup → register → activate → verify → mine → info → heartbeat ✅

+++ contracts/contracts/ComputeIndexToken.sol
@@ -33,8 +33,8 @@ interface IComputeRegistry {
         bool verified;
     }
 
-    function getNode(uint256 nodeId) external view returns (ComputeNode memory);
-    function addRevenue(uint256 nodeId, uint96 amount) external;
+    function getNode(uint64 nodeId) external view returns (ComputeNode memory);
+    function addRevenue(uint64 nodeId, uint96 amount) external;
 }
 
 interface IPriceOracle {
@@ -59,7 +59,7 @@ contract ComputeIndexToken {
     IPriceOracle public oracle;
 
     struct Deposit {
-        uint256 nodeId;        // which node backs this deposit
+        uint64 nodeId;          // which node backs this deposit
         uint256 amountWei;     // BOT deposited
         uint256 mintedTokens;  // CIF tokens minted
         uint64 depositedAt;
@@ -75,7 +75,7 @@ contract ComputeIndexToken {
     // ── Events ─────────────────────────────────────────────
     event Transfer(address indexed from, address indexed to, uint256 value);
     event Approval(address indexed owner, address indexed spender, uint256 value);
-    event Deposited(address indexed provider, uint256 indexed nodeId, uint256 amountWei, uint256 mintedTokens);
+    event Deposited(address indexed provider, uint64 indexed nodeId, uint256 amountWei, uint256 mintedTokens);
     event Withdrawn(address indexed provider, uint256 amountWei, uint256 burnedTokens);
 
     // ── Constructor ────────────────────────────────────────
@@ -92,7 +92,7 @@ contract ComputeIndexToken {
      *         Mint amount = deposit * issuanceRatio (based on node revenue & AI valuation).
      * @param nodeId The compute node backing this deposit
      */
-    function depositRevenue(uint256 nodeId) external payable returns (uint256 minted) {
+    function depositRevenue(uint64 nodeId) external payable returns (uint256 minted) {
         require(msg.value > 0, "Must deposit BOT");
 
         IComputeRegistry.ComputeNode memory node = registry.getNode(nodeId);
